+++ src/Interfaces/Libs/Deal.php
@@ -6,11 +6,11 @@ use \Hydra\Exchange\Interfaces\Entities\Order as iOrder;
 
 interface Deal
 {
-    public function getPrice() : int;
+    public function getPrice() : float;
 
     public function getType() : int;
 
-    public function getQuantity() : int;
+    public function getQuantity() : float;
 
     public function getBuyOrder() : iOrder;
